give bot messages priority over reply_broadcast when interpreting messages
we had a rather cryptic error

```
 error in process sentinel: Unbound slot: slack-reply-broadcast-message, "#<slack-reply-broadcast-message slack-reply-broadcast-message-57d04eb2>", user oref ```

which was trying to tell us that reply_broadcast messages come without
an user id. In truth only apps that send messages to a thread produce
this kind of reply_broadcast without user id. They have a bot_id so
they should be treated as bot messages. So I switched the order of our
message interpretation.