make sure faces shows for the markdown text in sent messages
at least in my system when I sent a message with ~striked text~, it
would show it as plain text (making me unsure if the striking
succeded). Apparently font-lock was messing up with the faces.
I am not sure if other places need this change, I will apply if I find
it out.