(mostly) fix user presence awareness
The user presence is not shared automatically anymore.
The simplistic approach taken here is to subscribe and query the first
499 users, giving priority to the ones you have private conversations
open with.

A slightly better solution would be to also proritize users that are
present in open channels for subscription. Possibly we can query more
thant 499 by batching, but there would be a delay of few seconds in
between web-socket reconnects.