add custom notifications
Sometimes you want to receive a notification for a channel you are not
subscribed to. For example, I have a channel that tells when
software deployments are completed and sometimes I want to react to
those messages: now I can just extend the notification system to send
me a message if the name of the software being deployed appears in
that channel.

Potentially, this could be extended to run actions when a condition is
met, but I didn't look into that yet.

You can select the message (with a string contained
in the message text) you want to be notified
for with `slack-message-add-text-custom-notification-predicate`. This
will notify only for a single message.
`slack-message-add-persistent-text-custom-notification-predicate` will
notifiy for the rest of the session.