support again remind for later and clean up slack-reminder.el
the reminder API is dropped in favour of the (unpublished) saved.add.
So here I remove the old functionality and support reminder via
saved.add.

Work to do is to show the due dates when listing the reminders and a
way to mark them done (that should be easy as there is an update for
that API), for now they can just be deleted.