allow to open conversations when the room wasn't yet opened
sometimes you want to open a conversation with a new group of people.
emacs-slack has failed to do that for too long. The issue was that we
were just hoping the room existed in our cache. Instead in the above
case we are creating the room for the first time, so we need to fetch
and store its information.

This commit adds that functionality for user profiles direct messages,
im and group messages.