fix regression on downloading default emojis
when I "fixed" the request wrapper to actually run requests
synchronously and return the request object, I didn't see there was
still some usage with `:sync t` and a success/error callback.

I saw that in case of sync request, the callback makes the request
fail, so I dropped it in this commit.

By the way, it seemed that slack-emoji-master-image-url also stopped
working to fetch the emoji png, so fixed it as well.

addresses https://github.com/emacs-slack/emacs-slack/issues/625