Render messages with attachment blocks
Some messages contain `attachments` with `blocks` field (instead of
directly having `blocks` field), which are just regular Slack
blocks. Here we check for them while rendering messages and if there
is an attachment with blocks, we also render it.

Not quite sure if it's possible to have more than one attachment and
they have blocks underneath them. I only saw one attachment with
multiple blocks. If there is such thing, we can extend here in the
future.